cogl: Fix signage miss match warnings
https://bugzilla.gnome.org/show_bug.cgi?id=768976

+++ cogl/cogl/cogl-bitmask.h
@@ -300,7 +300,7 @@ _cogl_bitmask_popcount_upto (const CoglBitmask *bitmask,
 {
   if (_cogl_bitmask_has_array (bitmask))
     return _cogl_bitmask_popcount_upto_in_array (bitmask, upto);
-  else if (upto >= COGL_BITMASK_MAX_DIRECT_BITS)
+  else if (upto >= (int) COGL_BITMASK_MAX_DIRECT_BITS)
     return _cogl_util_popcountl (_cogl_bitmask_to_bits (bitmask));
   else
     return _cogl_util_popcountl (_cogl_bitmask_to_bits (bitmask) &
